# docs/assets/api.yml
openapi: 3.0.0
info:
  title: Robinson API
  version: 0.1.0
  description: Robinson RESTful API
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  contact:
    name: Ryan McCartney
    url: https://ryan.mccartney.info/robinson
    email: hello@mccartney.info
servers:
  - url: http://localhost:80/api/
paths:
...
  /users:
    get:
      summary: Get a list of all users
      tags:
        - users
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    post:
      summary: Add a new user
      tags:
        - users
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  /users/current:
    get:
      summary: Get the current user
      tags:
        - users
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    put:
      summary: Update the current user
      tags:
        - users
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  "/users/{userId}":
    get:
      summary: Get a user by their ID
      tags:
        - users
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    put:
      summary: Update a user by their ID
      tags:
        - users
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    delete:
      summary: Delete a user by their ID
      tags:
        - users
      parameters:
        - in: path
          name: userId
          schema:
            type: string
          required: true
          description: The user ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
...